% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinkdays.R
\docType{data}
\name{drinkdays}
\alias{drinkdays}
\title{DRINKDAYS}
\format{
An object of class \code{data.frame} with 300 rows and 3 columns.
}
\usage{
data(drinkdays)
}
\description{
An example data set from simulated data.
}
\examples{
\dontrun{
data(drinkdays)
}
}
\keyword{datasets}
