% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_ggcorset}
\alias{theme_ggcorset}
\title{THEME_GGCORSET}
\usage{
theme_ggcorset()
}
\value{
ggplot2 theme
}
\description{
This function offers a ggplot theme to make visualizations more polished.
}
\examples{

wide.df <- data.frame(id = c(1,2,3,4,5),
             time1 = c(3,4,7,5,6),
             time2  = c(5,5,7,3,0),
             change = c(28.57,14.29,0,-28.57,-85.71))

plot1 <- gg_corset(data = wide.df, y_var1 = "time1", y_var2 = "time2",
          group = "id", c_var = "change")

plot1 + theme_ggcorset()

}
