% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_corset.R
\name{gg_corset}
\alias{gg_corset}
\title{CORSET PLOT}
\arguments{
\item{data}{The name of the data frame.}

\item{y_var1}{The name of measured variable at time 1.}

\item{y_var2}{The name of measured variable at time 2.}

\item{group}{The name of units measured at each time point such as 'ID'.}

\item{c_var}{The name of variable to visualize by line colour, such as percent change.}

\item{vio_fill}{The fill colour of the half violins. Optional (defaults to a soft black).}
}
\description{
This function visualizes a corset plot in wide format.
}
\examples{
\dontrun{
wide.df <- c(id = c(1,2,3,4,5),
             time1 = c(3,4,7,5,6),
             time2  = c(5,5,7,3,0)),
             change = c(28.57,14.29,0,-28.57,-85.71))
gg_corset(data = wide.df, y_var1 = "time1",
          y_var2 = "time2", group = "id", c_var = "change")
}
}
